~: improved code
