+++ package.json
@@ -20,13 +20,18 @@
     "homepage": "https://github.com/leoBazot/escape#readme",
     "dependencies": {
         "@babylonjs/loaders": "^5.54.0",
-        "babylonjs": "^5.54.0",
         "cannon": "^0.6.2",
         "clean-webpack-plugin": "^4.0.0"
     },
     "devDependencies": {
         "@babylonjs/core": "^5.54.0",
         "@babylonjs/inspector": "^5.54.0",
+        "@typescript-eslint/eslint-plugin": "^5.59.0",
+        "eslint": "^8.38.0",
+        "eslint-config-standard-with-typescript": "^34.0.1",
+        "eslint-plugin-import": "^2.27.5",
+        "eslint-plugin-n": "^15.7.0",
+        "eslint-plugin-promise": "^6.1.1",
         "html-webpack-plugin": "^5.5.0",
         "source-map-loader": "^4.0.1",
         "ts-loader": "^9.4.2",
